README.md
This method is used to find the root of a given function given two initial points. The inputs are **func**- the function being evaluated, **x1**- the lower guess, **xu**- the upper guess, **es**-the desired relative error( .0001%) and **maxiter**-the number of iterations it should default to 200. The outputs are  **root**-the estimated root, **fx**-the function evaluated at the root, **ea**- the relative error, **iter**- the number of iterations.