ReadMe.md
This algorithm is used to solve sets of equations using the LU factorization method. The inputs are **A** -coefficient matrix,  the outputs are **L**- lower triangular matrix , **U**- upper triangular matrix and **P**- the pivot matrix. The limitation of this algorithm is that it does not contain appropriate error checking.