README.MD
This algorithm is used to integrate given a set of date. It uses the 1/3 rule as well as the trapezoidal rule if its needed to use on the last interval. The inputs are  **x**- a vector and  **y** -the outputs of function. The outputs are **integration**- the sum of the trapezoidal rule and the 1/3 rule.